Add PyCRAM tf_broadcaster to demo
To use the frames published by the PyCRAM tf_broadcaster the frames
are forwarded to the DAGAP service. The robot tf_root is uploaded
to the param server.
Also made changes in the DAGAP service to use the new frames published
by PyCRAM.
This solution probably makes the custom tf_broadcaster in DAGAP obsolete.
Will remove in a later commit.